% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align_sequences}
\alias{align_sequences}
\title{Align sequences}
\usage{
align_sequences(
  x,
  method = c("auto", "msa", "decipher", "mafft", "muscle", "none"),
  gap_threshold = 0.9,
  ...
)
}
\arguments{
\item{x}{A `vs_sequences` object.}

\item{method}{Backend to use. `"auto"` picks the first available of `msa`,
`decipher`, `mafft`, `muscle`, falling back to `"none"` if the input is
already aligned.}

\item{gap_threshold}{Drop alignment columns that are gaps in more than this
proportion of sequences. `NULL` disables trimming.}

\item{...}{Passed to the underlying aligner.}
}
\value{
A `vs_alignment` object wrapping a `phyDat` alignment plus metadata.
}
\description{
Align sequences
}
