% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-lrt.R
\name{anova_table}
\alias{anova_table}
\title{Analysis of variance table for the final model}
\usage{
anova_table(selection)
}
\arguments{
\item{selection}{A `vs_selection` object.}
}
\value{
A tibble of sequential (Type I) tests.
}
\description{
Analysis of variance table for the final model
}
