% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitution-model.R
\name{candidate_models}
\alias{candidate_models}
\title{Candidate substitution models}
\usage{
candidate_models(molecule = c("protein", "nucleotide"))
}
\arguments{
\item{molecule}{`"protein"` or `"nucleotide"`.}
}
\value{
Character vector of model names understood by \pkg{phangorn}.
}
\description{
Candidate substitution models
}
