% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-forward.R
\name{model_fit_summary}
\alias{model_fit_summary}
\title{Fit statistics for a selected model}
\usage{
model_fit_summary(model)
}
\arguments{
\item{model}{A fitted model.}
}
\value{
A list of fit statistics.
}
\description{
For mixed models, marginal and conditional pseudo R-squared are reported in
place of adjusted R-squared.
}
