% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{pairwise_distances}
\alias{pairwise_distances}
\title{Pairwise distance matrix for an alignment}
\usage{
pairwise_distances(alignment, metric = "ml_aa", model = NULL, ...)
}
\arguments{
\item{alignment}{A `vs_alignment` object.}

\item{metric}{One of [distance_metrics()].}

\item{model}{Substitution model for the ML metrics. Accepts a
`vs_model_selection` object, in which case its chosen model is used.}

\item{...}{Passed to the underlying distance function.}
}
\value{
A `dist` object.
}
\description{
Pairwise distance matrix for an alignment
}
