% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_distances}
\alias{plot_distances}
\title{Distribution of divergence from consensus}
\usage{
plot_distances(x, by = "country")
}
\arguments{
\item{x}{A `vs_dataset`, `vs_distances` or `vs_analysis`.}

\item{by}{Optional grouping column, e.g. `"country"` or `"host_class"`.}
}
\value{
A ggplot object.
}
\description{
Distribution of divergence from consensus
}
