% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_gvif}
\alias{plot_gvif}
\title{Collinearity diagnostics}
\usage{
plot_gvif(x, threshold = 10)
}
\arguments{
\item{x}{A fitted model, `vs_selection` or `vs_analysis`.}

\item{threshold}{Threshold line to draw.}
}
\value{
A ggplot object.
}
\description{
Collinearity diagnostics
}
