% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_model_selection}
\alias{plot_model_selection}
\title{Substitution model ranking}
\usage{
plot_model_selection(x)
}
\arguments{
\item{x}{A `vs_model_selection` or `vs_analysis`.}
}
\value{
A ggplot object.
}
\description{
Substitution model ranking
}
