% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_predictor}
\alias{plot_predictor}
\title{Relationship between a predictor and divergence}
\usage{
plot_predictor(x, predictor, colour_by = "country")
}
\arguments{
\item{x}{A `vs_dataset` or `vs_analysis`.}

\item{predictor}{Column name of the predictor.}

\item{colour_by}{Optional grouping column for point colour.}
}
\value{
A ggplot object.
}
\description{
Relationship between a predictor and divergence
}
