% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_selection_path}
\alias{plot_selection_path}
\title{Information criterion path across forward selection}
\usage{
plot_selection_path(selection)
}
\arguments{
\item{selection}{A `vs_selection` or `vs_analysis`.}
}
\value{
A ggplot object.
}
\description{
Information criterion path across forward selection
}
