% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-environment.R
\name{read_environment}
\alias{read_environment}
\title{Read environmental data from a local file}
\usage{
read_environment(path, country_col = "country", year_col = "year", ...)
}
\arguments{
\item{path}{Path to a CSV.}

\item{country_col, year_col}{Names of the country and year columns.}

\item{...}{Passed to [utils::read.csv()].}
}
\value{
A tibble with `country` and `year` normalised.
}
\description{
Use this when you have compiled predictors yourself, for example tree cover
loss from Global Forest Watch or land cover from the ESA CCI, neither of
which has an open API. Subnational data is supported: supply an `admin1`
column and pass `by = c("country", "admin1", "year")` to [join_environment()].
}
