% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sea_countries}
\alias{sea_countries}
\title{Standard ISO3 codes for the countries used in the source study}
\usage{
sea_countries()
}
\value{
A named character vector mapping country name to ISO3 code.
}
\description{
Standard ISO3 codes for the countries used in the source study
}
