% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{viroscape-package}
\alias{viroscape}
\alias{viroscape-package}
\title{viroscape: Ecological Predictors of Viral Genetic Divergence}
\description{
A reproducible platform for quantifying the relationship between ecological, land-use and climate variables and viral genetic divergence. Implements the full pipeline used in Au & Buchan (2026) for H5Nx haemagglutinin in Southeast Asia: sequence retrieval, alignment, consensus construction, substitution-model selection by information criterion, divergence-from-consensus at amino acid or nucleotide level, environmental data joining, and forward stepwise regression with collinearity and coefficient-stability guards. Includes a Shiny front-end for interactively swapping regions, distance metrics, substitution models and predictor sets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frederickau/viroscape}
}

}
\author{
\strong{Maintainer}: Frederick Au \email{frederickau87@gmail.com}

Authors:
\itemize{
  \item Frederick Au \email{frederickau87@gmail.com}
}

Other contributors:
\itemize{
  \item Samantha Buchan [contributor]
}

}
\keyword{internal}
