% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{vs_cache_list}
\alias{vs_cache_list}
\alias{vs_cache_clear}
\title{List or clear cached downloads}
\usage{
vs_cache_list(cache_dir = vs_cache_dir())

vs_cache_clear(cache_dir = vs_cache_dir())
}
\arguments{
\item{cache_dir}{Cache directory.}
}
\value{
For `vs_cache_list()`, a tibble of cache entries. `vs_cache_clear()`
  returns the number of files removed, invisibly.
}
\description{
List or clear cached downloads
}
